Moved commands into their own directory
* Broke command classes out into their own files.  The commands file
 was quite large, and breaking out into individual files felt like the
 right thing to do.
* Added missing tests.

Fixes: #256
