+++ molecule/cli.py
@@ -30,7 +30,7 @@ Commands:
     verify        create, provision and test instances
     destroy       destroy instances
     status        show status of instances
-    list          show available platforms, providers
+    list          show available platforms
     login         connects to instance via SSH
     init          creates the directory structure and files for a new Ansible role compatible with molecule
 
@@ -39,12 +39,11 @@ Options:
     -v --version  shows the version
 """
 
-import sys
-
 import docopt
 
 import molecule
 import molecule.commands
+from molecule import utilities
 
 
 class CLI(object):
@@ -52,16 +51,17 @@ class CLI(object):
         args = docopt.docopt(__doc__,
                              version=molecule.__version__,
                              options_first=True)
-        command_name = args.get('<command>').capitalize()
+        command_name = args.get('<command>')
         command_args = {} if args.get('<args>') is None else args.pop('<args>')
 
         try:
-            command_class = getattr(molecule.commands, command_name)
+            command_module = getattr(molecule.commands, command_name)
+            command_clazz = getattr(command_module, command_name.capitalize())
         except AttributeError:
             raise docopt.DocoptExit()
 
-        c = command_class(command_args, args)
-        sys.exit(c.execute()[0])
+        c = command_clazz(command_args, args)
+        utilities.sysexit(c.execute()[0])
 
 
 def main():
