+++ molecule/utilities.py
@@ -77,6 +77,7 @@ def merge_dicts(a, b, raise_conflicts=False, path=None):
     """
     Merges the values of B into A.
     If the raise_conflicts flag is set to True, a LookupError will be raised if the keys are conflicting.
+
     :param a: the target dictionary
     :param b: the dictionary to import
     :param raise_conflicts: flag to raise an exception if two keys are colliding
@@ -120,6 +121,7 @@ def merge_dicts(a, b, raise_conflicts=False, path=None):
 def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
     """
     Writes a file from a jinja2 template.
+
     :param src: the target template files to use
     :param dest: destination of the templatized file to be written
     :param kwargs: dictionary of arguments passed to jinja2 when rendering template
@@ -133,8 +135,7 @@ def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
 
     # template file doesn't exist
     if path and not os.path.isfile(src):
-        logger.error('\n{}Unable to locate template file: {}{}'.format(
-            colorama.Fore.RED, src, colorama.Fore.RESET))
+        logger.error('\nUnable to locate template file: {}\n'.format(src))
         sys.exit(1)
 
     # look for template in filesystem, then molecule package
@@ -152,6 +153,7 @@ def write_template(src, dest, kwargs={}, _module='molecule', _dir='templates'):
 def write_file(filename, content):
     """
     Writes a file with the given filename using the given content. Overwrites, does not append.
+
     :param filename: the target filename
     :param content: what gets written into the file
     :return: None
@@ -163,6 +165,7 @@ def write_file(filename, content):
 def format_instance_name(name, platform, instances):
     """
     Takes an instance name and formats it according to options specified in the instance's config.
+
     :param name: the name of the instance
     :param platform: the current molecule platform in use
     :param instances: the current molecule instances dict in use
@@ -245,6 +248,7 @@ def check_ssh_availability(hostip, user, timeout):
 def debug(title, data):
     """
     Prints colorized output for use when debugging portions of molecule.
+
     :param title: title of debug output
     :param data: data of debug output
     :return: None
@@ -254,3 +258,7 @@ def debug(title, data):
                    colorama.Back.RESET, colorama.Style.RESET_ALL]))
     print(''.join([colorama.Fore.BLACK, colorama.Style.BRIGHT, data,
                    colorama.Style.RESET_ALL, colorama.Fore.RESET]))
+
+
+def sysexit(code=1):
+    sys.exit(code)
